fix warnings
git-svn-id: https://www.opensc-project.org/svnp/opensc/trunk@3004 c6295689-39f2-0310-b995-f0e70906c6a9

+++ src/tools/opensc-explorer.c
@@ -175,7 +175,7 @@ static void print_file(const sc_file_t *file)
 		break;
 	}
 	printf("\t%4s", st);
-	printf(" %5lu", file->size);
+	printf(" %5lu", (unsigned long)file->size);
 	if (file->namelen) {
 		printf("\tName: ");
 		print_binary(stdout, file->name, file->namelen);
